+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.0.2] - 2023-01-18
+
+### Added
+
+- `VvCombobox` support for floating label, unselectable items and native select;
+- `VvSelect` support for floating label, unselectable items, and multiple selection;
+- `VvDropdown` support for unselectable items;
+- `VvInputText` autofocus is triggered on element visibility;
+- `VvTextarea` autofocus is triggered on element visibility;
+- `VvSelect` autofocus is triggered on element visibility.
+
+### Fixed
+
+- `VvButtonGroup` itemModifiers now are merged with the child modifiers;
+- `VvInputText` and `VvTextarea` count without maxlength;
+- unplugin resolver fix for `VvCombobox`.
+
 ## [0.0.1] - 2023-01-17
 
 ### Added
@@ -26,3 +43,5 @@ All notable changes to this project will be documented in this file.
 - `VvSelect` component;
 - `VvTextarea` component;
 - `VvRadioGroup` component.
+
+[0.1.2]: https://github.com/volverjs/style/compare/v0.0.1...v0.0.2